+++ compilertools/compilers/llvm.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-"""GNU Compiler Collection"""
-# https://gcc.gnu.org/onlinedocs/gcc/Invoking-GCC.html
+"""LLVM Clang"""
+# https://clang.llvm.org/docs/UsersManual.html
 
 
 from compilertools.compilers import CompilerBase as _CompilerBase
@@ -9,7 +9,7 @@ __all__ = ['Compiler']
 
 
 class Compiler(_CompilerBase):
-    """GNU Compiler Collection"""
+    """LLVM Clang"""
 
     @_CompilerBase._memoized_property
     def option(self):
@@ -32,16 +32,10 @@ class Compiler(_CompilerBase):
             Keys are API names, values are dict of arguments
             with keys in {'link', 'compile'}."""
         api = {}
-        if self.version >= 4.2:
+        if self.version >= 3.7:
             api['openmp'] = {
                 'compile': '-fopenmp',
-                'link': '-fopenmp'}
-        if self.version >= 4.9:
-            api['cilkplus'] = {
-                'compile': '-fcilkplus -lcilkrts',
-                'link': '-fcilkplus -lcilkrts'}
-        if self.version >= 6.1:
-            api['openacc'] = {'compile': '-fopenacc'}
+                'link': '-fopenmp=libomp'}
         return api
 
     @_CompilerBase._memoized_property
@@ -57,16 +51,17 @@ class Compiler(_CompilerBase):
 
         from subprocess import Popen, PIPE
         try:
-            version_str = Popen(['gcc', '--version'], stdout=PIPE,
+            version_str = Popen(['clang', '--version'], stdout=PIPE,
                                 universal_newlines=True).stdout.read()
         except OSError:
             return
 
-        if 'gcc' not in version_str.lower():
-            # "gcc" command is linked to non GCC compiler on some systems
+        if 'clang' not in version_str.lower():
+            # "clang" command seem to not run CLang
             return
 
-        version_str = version_str.split(')', 1)[1].rstrip().split(maxsplit=1)[0]
+        version_str = version_str.split('(', 1)[0]
+        version_str = version_str.split('version', 1)[1].strip()
 
         # Keep only major and minor
         return float(version_str.rsplit('.', 1)[0])
@@ -82,7 +77,7 @@ class Compiler(_CompilerBase):
         from platform import python_compiler
         version_str = python_compiler()
 
-        if 'GCC' not in version_str:
+        if 'clang' not in version_str.lower():
             return 0.0
 
         version_str = version_str.split(' ', 2)[1]
@@ -91,7 +86,7 @@ class Compiler(_CompilerBase):
         return float(version_str.rsplit('.', 1)[0])
 
     def _compile_args_matrix(self, arch, cpu):
-        """Returns available GCC compiler options for the
+        """Returns available Clang compiler options for the
         specified CPU architecture.
 
         Parameters
@@ -120,18 +115,16 @@ class Compiler(_CompilerBase):
                           import_if=('AVX512F' in cpu.features and
                                      'AVX512CD' in cpu.features and
                                      cpu.os_supports_xsave),
-                          build_if=self.version >= 4.9),
+                          build_if=self.version >= 3.9),
 
                  self.Arg(args='-mavx2',
                           suffix='avx2',
-                          import_if=(self.version >= 4.7 and
-                                     'AVX2' in cpu.features and
+                          import_if=('AVX2' in cpu.features and
                                      cpu.os_supports_xsave)),
 
                  self.Arg(args='-mavx',
                           suffix='avx',
-                          import_if=(self.version >= 4.4 and
-                                     'AVX' in cpu.features and
+                          import_if=('AVX' in cpu.features and
                                      cpu.os_supports_xsave)),
                  self.Arg(),
                  ],
@@ -139,8 +132,7 @@ class Compiler(_CompilerBase):
                 # CPU Generic vendor/brand optimisations
                 [self.Arg(args='-mtune=intel',
                           suffix='intel',
-                          import_if=cpu.vendor == 'GenuineIntel',
-                          build_if=self.version >= 4.9),
+                          import_if=cpu.vendor == 'GenuineIntel'),
 
                  self.Arg(),
                  ]
@@ -154,52 +146,43 @@ class Compiler(_CompilerBase):
                 # CPU Instructions sets
                 [self.Arg(args=['-mfpmath=sse', '-mavx2'],
                           suffix='avx2',
-                          import_if=(self.version >= 4.7 and
-                                     'AVX2' in cpu.features and
+                          import_if=('AVX2' in cpu.features and
                                      cpu.os_supports_xsave)),
 
                  self.Arg(args=['-mfpmath=sse', '-mavx'],
                           suffix='avx',
-                          import_if=(self.version >= 4.4 and
-                                     'AVX' in cpu.features and
+                          import_if=('AVX' in cpu.features and
                                      cpu.os_supports_xsave)),
 
                  self.Arg(args=['-mfpmath=sse', '-msse4'],
                           suffix='sse4',
                           import_if=('SSE4_1' in cpu.features and
-                                     'SSE4_2' in cpu.features),
-                          build_if=self.version >= 4.3),
+                                     'SSE4_2' in cpu.features)),
 
                  self.Arg(args=['-mfpmath=sse', '-msse4.2'],
                           suffix='sse4_2',
-                          import_if='SSE4_2' in cpu.features,
-                          build_if=self.version >= 4.3),
+                          import_if='SSE4_2' in cpu.features),
 
                  self.Arg(args=['-mfpmath=sse', '-msse4.1'],
                           suffix='sse4_1',
-                          import_if='SSE4_1' in cpu.features,
-                          build_if=self.version >= 4.3),
+                          import_if='SSE4_1' in cpu.features),
 
                  self.Arg(args=['-mfpmath=sse', '-msse4a'],
                           suffix='sse4a',
-                          import_if=(self.version >= 4.9 and
-                                     'SSE4A' in cpu.features and
+                          import_if=('SSE4A' in cpu.features and
                                      cpu.vendor == 'AuthenticAMD')),
 
                  self.Arg(args=['-mfpmath=sse', '-mssse3'],
                           suffix='ssse3',
-                          import_if='SSSE3' in cpu.features,
-                          build_if=self.version >= 4.3),
+                          import_if='SSSE3' in cpu.features),
 
                  self.Arg(args=['-mfpmath=sse', '-msse2'],
                           suffix='sse2',
-                          import_if='SSE2' in cpu.features,
-                          build_if=self.version >= 3.3),
+                          import_if='SSE2' in cpu.features),
 
                  self.Arg(args=['-mfpmath=sse', '-msse'],
                           suffix='sse',
-                          import_if='SSE' in cpu.features,
-                          build_if=self.version >= 3.1),
+                          import_if='SSE' in cpu.features),
 
                  self.Arg(),
                  ],
@@ -207,8 +190,7 @@ class Compiler(_CompilerBase):
                 # CPU Generic vendor/brand optimisations
                 [self.Arg(args='-mtune=intel',
                           suffix='intel',
-                          import_if=cpu.vendor == 'GenuineIntel',
-                          build_if=self.version >= 4.9),
+                          import_if=cpu.vendor == 'GenuineIntel'),
 
                  self.Arg(),
                  ]
@@ -217,7 +199,7 @@ class Compiler(_CompilerBase):
         return args
 
     def _compile_args_current_machine(self, arch, cpu):
-        """Return auto-optimised GCC arguments for current machine.
+        """Return auto-optimised Clang arguments for current machine.
 
         Parameters
         ----------
@@ -236,7 +218,7 @@ class Compiler(_CompilerBase):
         # Arch specific optimizations
         if arch == 'x86_32':
             args.append('-m32')
-            if 'SSE' in cpu.features and self.version >= 3.1:
+            if 'SSE' in cpu.features:
                 args.append('-mfpmath=sse')
 
         elif arch == 'x86_64':